+++ environment_wbd.yml
@@ -9,6 +9,7 @@ dependencies:
   - geopandas>=0.13.1,<2.0
   - ipython>=8.29.0,<9.0.0
   - jupyter>=1.0
+  - lxml
   - matplotlib>=3.5,<4.0
   - mplleaflet>=0.0.5,<1.0
   - netcdf4>=1.6.4,<2.0
@@ -25,6 +26,7 @@ dependencies:
   - xlrd>=2.0.1,<3.0
   - momepy # TODO: kijken of we dit eruit kunnen halen
   - rioxarray=0.18.2
+  - ruff
   - geocube
   - networkx
   - pcraster
