+++ wbd_tools/process_damo/__init__.py
@@ -1,0 +1,8 @@
+from wbd_tools.process_damo.process_closings import ProcessClosings
+from wbd_tools.process_damo.process_culverts import ProcessCulverts
+from wbd_tools.process_damo.process_network import ProcessNetwork
+from wbd_tools.process_damo.process_profiles import ProcessProfiles
+from wbd_tools.process_damo.process_pumps import ProcessPumps
+from wbd_tools.process_damo.process_weirs import ProcessWeirs
+
+__all__ = ["ProcessClosing", "ProcessCulverts","ProcessNetwork","ProcessProfiles","ProcessPumps","ProcessWeirs"]
