+++ wbd_tools/tohydamogml/__init__.py
@@ -1,0 +1,2 @@
+from wbd_tools.tohydamogml.hydamo_table import HydamoObject
+__all__ = ["HydamoObject"]
